simplify shared memory model
The performance overhead of allocating shared memory buffers was
measured to be negligible. The shared memory is now allocated on each
frame and not reused. This means, image buffers do not get reused and
may be assumed to be immutable. This means, we need no further
synchronization or reference counting of buffers, since processes may
only read and can leave cleanup to the OS by calling close().