synchronize expected entities
Removing or adding entities from the last_entities field works at the
beginning of tracking, but it breaks if entities are added during
tracking: When a tracking task is running, it uses a copy of
last_entities for matching. After the task is finished, this in-flight
copy will be used to update last_entities, overwriting any user changes
made in the meantime. To fix this, expected entity ids are now stored in
a list, which is synchronized with last_entities before starting
tracking tasks.