refactor features
Features used to contain the skeleton nodes in image space as well as a
pose in world space. This convention assumes that plugins need the
skeleton nodes in image space, which may not always be the case.

This commit removes the world pose from the feature message. Instead,
the skeleton nodes are now stored twice, to represent both world and
image space. It is sufficient to emit nodes in image space during
feature detection, the core will automatically transform them to world
space. Storing the nodes twice consumes more memory, but it also
simplifies plugin development, since rectification and undistortion
is done by the core.

Poses are not stored anymore, they can be calculated from features on
the fly. This can be achieved in python by calling the function
`feature_to_world_pose`.